flask_docker_kidternship/requirements.txt
Flask
pytz
flask-cors
http.client
requests
flask-socketio
better_profanity
psycopg2
wordcloud
numpy
matplotlib
Pillow
nltk